enh: split case with sinlge file
