+++ spark/src/main/java/com/taosdata/java/DemoSubscribe.java
@@ -1,0 +1,159 @@
+package com.taosdata.java;
+
+
+import java.sql.PreparedStatement;
+import java.sql.Timestamp;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.taosdata.java.SparkTest.ResultBean;
+import com.taosdata.jdbc.TSDBDriver;
+import com.taosdata.jdbc.tmq.*;
+import com.taosdata.jdbc.utils.JsonUtil;
+
+import java.sql.*;
+import java.time.Duration;
+import java.util.Collections;
+import java.util.List;
+import java.util.Properties;
+import java.util.Set;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.Random;
+
+import org.apache.spark.sql.Dataset;
+import org.apache.spark.sql.DataFrameReader;
+import org.apache.spark.sql.Row;
+import org.apache.spark.sql.SaveMode;
+import org.apache.spark.sql.SparkSession;
+import org.apache.spark.sql.jdbc.JdbcDialect;
+import org.apache.spark.sql.jdbc.JdbcDialects;
+import org.java_websocket.framing.DataFrame;
+
+public class DemoSubscribe {
+
+    public static class ResultDeserializer extends ReferenceDeserializer<ResultBean> {}
+
+    // use this class to define the data structure of the result record
+    public static class ResultBean {
+        private Timestamp ts;
+        private double    current;
+        private int       voltage;
+        private double    phase;
+        private int       groupid;
+        private String    location;
+
+        public Timestamp getTs()             { return ts; }
+        public double    getCurrent()        { return current;}
+        public int       getVoltage()        { return voltage;}
+        public double    getPhase()          { return phase;}
+        public int       getGroupid()        { return groupid;}
+        public String    getLocation()       { return location;}
+
+        public void      setTs(Timestamp ts)          { this.ts       = ts;}
+        public void      setCurrent(double current)   { this.current  = current;}
+        public void      setVoltage(int voltage)      { this.voltage  = voltage;}
+        public void      setPhase(double phase)       { this.phase.   = phase;}
+        public void      setGroupid(int groupid)      { this.groupid  = groupid;}
+        public void      setLocation(String location) { this.location = location;}
+    }
+    
+    public static TaosConsumer<ResultBean> getConsumer() throws Exception {
+        // property
+        Properties config = new Properties();
+        config.setProperty("td.connect.type", "ws");
+        config.setProperty("bootstrap.servers", "localhost:6041");
+        config.setProperty("auto.offset.reset", "earliest");
+        config.setProperty("msg.with.table.name", "true");
+        config.setProperty("enable.auto.commit", "true");
+        config.setProperty("auto.commit.interval.ms", "1000");
+        config.setProperty("group.id", "group1");
+        config.setProperty("client.id", "clinet1");
+        config.setProperty("td.connect.user", "root");
+        config.setProperty("td.connect.pass", "taosdata");
+        config.setProperty("value.deserializer", "com.taosdata.java.SparkTest$ResultDeserializer");
+        config.setProperty("value.deserializer.encoding", "UTF-8");
+
+        try {
+            // new consumer
+            TaosConsumer<ResultBean> consumer= new TaosConsumer<>(config);
+            System.out.printf("Create consumer successfully, host: %s, groupId: %s, clientId: %s%n",
+                    config.getProperty("bootstrap.servers"),
+                    config.getProperty("group.id"),
+                    config.getProperty("client.id"));
+            return consumer;
+        } catch (Exception ex) {
+            // please refer to the JDBC specifications for detailed exceptions info
+            System.out.printf("Failed to create websocket consumer, host: %s, groupId: %s, clientId: %s, %sErrMessage: %s%n",
+                    config.getProperty("bootstrap.servers"),
+                    config.getProperty("group.id"),
+                    config.getProperty("client.id"),
+                    ex instanceof SQLException ? "ErrCode: " + ((SQLException) ex).getErrorCode() + ", " : "",
+                    ex.getMessage());
+            // Print stack trace for context in examples. Use logging in production.
+            ex.printStackTrace();
+            throw ex;
+        }
+    }
+
+    public static void pollExample(TaosConsumer<ResultBean> consumer) throws SQLException, JsonProcessingException {
+        List<String> topics = Collections.singletonList("topic_meters");
+        try {
+            // subscribe  topics
+            consumer.subscribe(topics);
+            System.out.println("Subscribe topics successfully.");
+            for (int i = 0; i < 100; i++) {
+                // poll data
+                ConsumerRecords<ResultBean> records = consumer.poll(Duration.ofMillis(100));
+                for (ConsumerRecord<ResultBean> record : records) {
+                    ResultBean bean = record.value();
+                    // Add your data processing logic here
+                    System.out.println("data: " + JsonUtil.getObjectMapper().writeValueAsString(bean));
+                }
+            }
+
+        } catch (Exception ex) {
+            // please refer to the JDBC specifications for detailed exceptions info
+            System.out.printf("Failed to poll data, topic: %s, groupId: %s, clientId: %s, %sErrMessage: %s%n",
+                    topics.get(0),
+                    groupId,
+                    clientId,
+                    ex instanceof SQLException ? "ErrCode: " + ((SQLException) ex).getErrorCode() + ", " : "",
+                    ex.getMessage());
+            // Print stack trace for context in examples. Use logging in production.
+            ex.printStackTrace();
+            throw ex;
+        }
+    }    
+
+    // subscribe
+    public static void subscribeFromTDengine() {
+        try {
+            // consumer
+            TaosConsumer<ResultBean> consumer = getConsumer();
+
+            // poll
+            pollExample(consumer);
+
+            // out succ
+            System.out.println("test subscribe successfully!");
+
+            // close
+            consumer.unsubscribe();
+            consumer.close();
+
+        } catch (SQLException ex) {
+            System.out.println("Failed to poll data from topic_meters, ErrCode:" + ex.getErrorCode() + "; ErrMessage: " + ex.getMessage());
+            return;
+        } catch (Exception ex) {
+            System.out.println("Failed to poll data from topic_meters, ErrMessage: " + ex.getMessage());
+            return;
+        }        
+    }
+}
